feat(runner): use tenv library to install binaries
